move worker solver to its own function
since clean with gdt cant use ProcessPool (because it cant
pickle nested functions), we need a better solver for the ammount
of workers depending on the type of multiprocess.

If ProcessPool, max workers is `os.cpu_count()`,
`extract_multiple` command uses this.

ThreadPool, max is `os.cpu_count() * 4`,
`clean` command uses this (right now, having gdt or not will run
with ThreadPool, but we might change to only use ThreadPool if it is
a gdt execution, depending on performance)

Both approachs returns the low value between user input and max.