added object pool to `orchestration()`, and standardized args in function calls
since a nested function is needed to support the option of
a gdt `.gdict` in `exec_single_an()` as `clean_func`, this means
that `exec_single_an` no longer can be call in ProcessPoolExecution.

Since it now needs ThreadPoolExecution, we can take advantage of
an object pool when handling log msgs. The changes made in `log_setup`
now supports both options of parallelism.