`_gdt_clean` is now pickleable, meaning no need for ThreadPool code
making _gdt_clean pickleable using functools partial, that way
we can use ProcessPoolExecutor, which for this application is
faster than ThreadPoolExecutor. The problem before was that all
args need to be pickleable to use Process, but since we were using
nested functions, we could not use it.

also clean up the code in clean.py, removing all the ThreadPool
shenanigans with Object Pool.

since the log is now also fully formatted in the main thread with
correct ordering, a lot of `[{an}]` has been removed from log msgs.