add GDT gdict as a separate server, where worker can request queries
change the way `multiple` process happens in `tigre clean`. Before gdict
as sent as another variable, however this became a bottleneck for large
gdicts, where every AN got a copy of the original gdict. Now we spawn
a dedicated `dictionary` process, that reads requests from a Queue,
and responds the `label` into a Pipe, created by the requesting worker.

This allows for great scallability, as the bottleneck is in the
server ability to clear the requests in the Queue. It is also easy
to scale horizontally, by spawning more dictionary servers and adopting
a round-robin approach when populating the request Queues.