better define `use_custom_builder`, by binding it with `types.MethodTypes`
this also changes the API for the creation of custom path function.

Now `use_custom_builder` will wrap and bind the first arg to `self`,
meaning that custom functions will have acess to all attrs that the
current instance has defined, buy using `self` as the first args in their
signature.
the example in the docstring shows that.